refactor out drug data parser
