+++ data-loader/src/v2/types/ParserOptions.ts
@@ -1,0 +1,11 @@
+export interface ParserOptions {
+  flags?: string;
+  encoding?: BufferEncoding;
+  fd?: number;
+  mode?: number;
+  autoClose?: boolean;
+  emitClose?: boolean;
+  start?: number;
+  end?: number;
+  highWaterMark?: number;
+}
