+++ data-loader/src/v2/types/index.ts
@@ -3,3 +3,4 @@ export * from './CSVData';
 export * from './CSVRow';
 export * from './EntityGraph';
 export * from './EntityNode';
+export * from './ParserOptions';
